refactor(test): split TestFixtures SoT and add 250-line ratchet (#4547)
Extract world-state, game-core, domain-game, unit, and combat factories under
lib/src/ with a thin game_test_fixtures.dart tear-off facade. Route all Game
and WorldState construction through single seams; add coverage for combatGame
and unit helpers; register repo.colonizethis_test_* file-size gates.
