+++ SPEC/program/repo-lint.md
@@ -267,6 +267,8 @@
 | `tool/check_economy_test_mirrors_lib.dart` | Keep the `packages/colonizethis_economy/test/**` tree mirrored to `lib/src/economy/` per the testing rule (`test/` mirrors `lib/`): any Dart source that sits **directly** at the `test/` root is a violation — both `*_test.dart` files and co-located support helpers must live in a subdirectory such as `test/economy/` or `test/economy/world_market/` that mirrors `lib/src/economy/`. Complements the testing-rule policy (Refs #4299); rule `repo.economy_test_mirrors_lib` |
 | `tool/check_save_source_file_size.dart` | PR-blocking physical-line gate: every `.dart` file under `packages/colonizethis_save/lib/src/**` must stay at or below **400** physical lines (empty shrink-only allowlist); keeps the #4077 `GameSaveAdapter` concern splits from re-merging; rule `repo.save_source_file_size` (Refs #4077) |
 | `tool/check_save_test_file_size.dart` | PR-blocking physical-line gate: every `.dart` file under `packages/colonizethis_save/test/**` must stay at or below **400** physical lines (empty shrink-only allowlist; mirrors `repo.economy_test_file_size` / `repo.logic_test_file_size`); rule `repo.save_test_file_size` (Refs #4077) |
+| `tool/check_colonizethis_test_lib_file_size.dart` | PR-blocking physical-line gate: every `.dart` file under `packages/colonizethis_test/lib/**` must stay at or below **250** physical lines (empty shrink-only allowlist); rule `repo.colonizethis_test_lib_file_size` (Refs #4547) |
+| `tool/check_colonizethis_test_test_file_size.dart` | PR-blocking physical-line gate: every `.dart` file under `packages/colonizethis_test/test/**` must stay at or below **250** physical lines (empty shrink-only allowlist); rule `repo.colonizethis_test_test_file_size` (Refs #4547) |
 | `tool/check_economy_test_no_local_run_scenarios.dart` | Forbid local `void Function() run` scenario shells under `packages/colonizethis_economy/test/**`; pins live in `colonizethis_economy_test_support` (Refs #4014); rule `repo.economy_test_no_local_run_scenarios` |
 | `tool/check_app_test_file_size.dart` | PR-blocking physical-line gate: `app/test/**/*.dart` must stay at or below **400** lines; `appTestFileSizeAllowlistForTests` is **empty** after wave-15 Slice D densify (stale/missing entries still fail; Refs #4013, #4021, #4305, #4352); rule `repo.app_test_file_size`. Distinct from repo-wide `repo.dart_file_non_comment_line_size` (1000 **NCL**). |
 | `tool/check_app_e2e_support_lib_file_size.dart` | PR-blocking physical-line gate: `packages/colonizethis_app_e2e_support/lib/**/*.dart` must stay at or below **300** lines; oversized files live on a **shrink-only** allowlist (`appE2eSupportLibFileSizeAllowlistForTests`) — missing or now-under-cap entries fail as stale (Refs #4075, #4195, #4344); rule `repo.app_e2e_support_lib_file_size`. |
@@ -732,6 +734,10 @@ Do **not** add new top-level `tool/check_*.dart` **entrypoints** for CI without 
 - Given a `packages/colonizethis_save/lib/src/**` Dart file with more than 400 physical lines, when `dart run tool/ct_repo_lint.dart` runs rule `repo.save_source_file_size`, then the run fails naming that file and exits `1` (Refs #4077).
 - Given every `.dart` file under `packages/colonizethis_save/test/**` is at or below 400 physical lines, when `dart run tool/ct_repo_lint.dart` runs rule `repo.save_test_file_size`, then the rule passes and exits `0` (Refs #4077).
 - Given a `packages/colonizethis_save/test/**` Dart file with more than 400 physical lines, when `dart run tool/ct_repo_lint.dart` runs rule `repo.save_test_file_size`, then the run fails naming that file and exits `1` (Refs #4077).
+- Given every `.dart` file under `packages/colonizethis_test/lib/**` is at or below 250 physical lines, when `dart run tool/ct_repo_lint.dart` runs rule `repo.colonizethis_test_lib_file_size`, then the rule passes and exits `0` (Refs #4547).
+- Given a `packages/colonizethis_test/lib/**` Dart file with more than 250 physical lines, when `dart run tool/ct_repo_lint.dart` runs rule `repo.colonizethis_test_lib_file_size`, then the run fails naming that file and exits `1` (Refs #4547).
+- Given every `.dart` file under `packages/colonizethis_test/test/**` is at or below 250 physical lines, when `dart run tool/ct_repo_lint.dart` runs rule `repo.colonizethis_test_test_file_size`, then the rule passes and exits `0` (Refs #4547).
+- Given a `packages/colonizethis_test/test/**` Dart file with more than 250 physical lines, when `dart run tool/ct_repo_lint.dart` runs rule `repo.colonizethis_test_test_file_size`, then the run fails naming that file and exits `1` (Refs #4547).
 - Given every non-generated `.dart` file under `packages/colonizethis_debug_console/lib/**` is at or below 270 physical lines, when `dart run tool/ct_repo_lint.dart` runs rule `repo.debug_console_lib_file_size`, then the rule passes and exits `0` (Refs #4433).
 - Given a `packages/colonizethis_debug_console/lib/**` Dart file with more than 270 physical lines, when `dart run tool/ct_repo_lint.dart` runs rule `repo.debug_console_lib_file_size`, then the run fails naming that file and exits `1` (Refs #4433).
 - Given every `.dart` file under `packages/colonizethis_debug_console/test/**` is at or below 400 physical lines, when `dart run tool/ct_repo_lint.dart` runs rule `repo.debug_console_test_file_size`, then the rule passes and exits `0` (Refs #4433).
